Add progress argument to homology_persistence() (fixes #19)

+++ include/dionysus/dlog/progress.h
@@ -11,11 +11,11 @@ namespace dlog
 
 struct progress
 {
-                        progress(unsigned total):
+                        progress(size_t total):
                             current_(0), total_(total)      { show_progress(); }
 
     progress&           operator++()                        { current_++;       if (current_ * 100 / total_ > (current_ - 1) * 100 / total_) show_progress(); check_done(); return *this; }
-    progress&           operator=(unsigned cur)             { current_ = cur;   show_progress(); check_done(); return *this; }
+    progress&           operator=(size_t cur)               { current_ = cur;   show_progress(); check_done(); return *this; }
     progress&           operator()(const std::string& s)    { message_ = s;     show_progress(); check_done(); return *this; }
     template<class T>
     progress&           operator()(const T& x)              { std::ostringstream oss; oss << x; return (*this)(oss.str()); }
@@ -24,7 +24,7 @@ struct progress
     void                check_done() const                  { if (current_ >= total_) std::cout << "\n" << std::flush; }
 
     private:
-        unsigned        current_, total_;
+        size_t          current_, total_;
         std::string     message_;
 };
